refactor:Component configuration and routing modify
