+++ src/components/presentational/group/GroupDetail.vue
