+++ src/pages/group/_id.vue
@@ -1,0 +1,9 @@
+<template>
+  <router-view />
+</template>
+
+<script>
+export default {
+  name: 'GroupPage'
+}
+</script>
