+++ src/pages/group/list.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script>
-import GroupListContainer from '@/components/container/groups/GroupListContainer.vue'
+import GroupListContainer from '@/components/container/group/GroupListContainer.vue'
 export default {
   name: 'GroupsPage',
   components: {
