+++ src/router/index.ts
@@ -1,7 +1,8 @@
 import Vue from 'vue'
 import VueRouter from 'vue-router'
-import groups from '@/pages/groups/index.vue'
-import groupDetail from '@/pages/groups/_id.vue'
+import groups from '@/pages/group/list.vue'
+import groupId from '@/pages/group/_id.vue'
+import groupDetail from '@/pages/group/detail.vue'
 
 Vue.use(VueRouter)
 
@@ -18,8 +19,17 @@ const routes = [
   },
   {
     path: '/groups/:id',
-    name: 'groupDetail',
-    component: groupDetail
+    component: groupId,
+    children: [
+      {
+        path: '',
+        redirect: '/groups/:id/detail'
+      },
+      {
+        path: 'detail',
+        component: groupDetail
+      }
+    ]
   },
   {
     path: '*',
